chore: codebase improvements (#808)
* chore: add ChainMonitor to Shasta node

* chore: rm param

* feat: change type of Proposal.num_forced_inclusion to u16

+++ shasta/src/node/proposal_manager/proposal.rs
@@ -19,7 +19,7 @@ pub struct Proposal {
     pub anchor_block_timestamp_sec: u64,
     pub anchor_block_hash: B256,
     pub anchor_state_root: B256,
-    pub num_forced_inclusion: u8,
+    pub num_forced_inclusion: u16,
 }
 
 impl Proposal {
